@@ -13,14 +13,14 @@
     "@fluentui/react-make-styles": "^9.0.0-beta.4",
     "@nextui-org/react": "^1.0.6-beta.5",
     "@tauri-apps/api": "^1.0.0-rc.3",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "use-dark-mode": "^2.3.1"
   },
   "devDependencies": {
     "@tauri-apps/cli": "^1.0.0-rc.8",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
     "@typescript-eslint/eslint-plugin": "^5.19.0",
     "@typescript-eslint/parser": "^5.19.0",
     "@vitejs/plugin-react": "^1.0.7",
